+++ KoiCareSystem/KoiCareSystem.Data/Models/OrderItem.cs
@@ -7,15 +7,15 @@ namespace KoiCareSystem.Data.Models;
 
 public partial class OrderItem
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
 
-    public long OrderId { get; set; }
+    public int OrderId { get; set; }
 
-    public long ProductId { get; set; }
+    public int ProductId { get; set; }
 
-    public long Quantity { get; set; }
+    public int Quantity { get; set; }
 
-    public long Price { get; set; }
+    public decimal Price { get; set; }
 
     public virtual Order Order { get; set; }
